Add files via upload
This project involves predicting employee salaries using a Random Forest Regressor model based on features like gender, education, age, years of experience, and job title. The model is trained on a dataset, and SHAP (Shapley Additive Explanations) is used to interpret the predictions by explaining how each feature influences the salary prediction. A key part of the project is modifying the Years of Experience feature and observing how it affects the salary prediction. The project provides both accurate salary predictions and transparent model explanations, helping to understand the contribution of each input feature.